fix : Product, Payment, Inventory Domain Package 구조 수정
