--- src/main/java/kr/co/popoolserver/inventory/controller/InventoryController.java
@@ -1,4 +1,0 @@
-package kr.co.popoolserver.inventory.controller;
-
-public class InventoryController {
-}
