--- src/main/java/kr/co/popoolserver/inventory/domain/InventoryDto.java
@@ -1,4 +1,0 @@
-package kr.co.popoolserver.inventory.domain;
-
-public class InventoryDto {
-}
