--- src/main/java/kr/co/popoolserver/inventory/domain/service/InventoryService.java
@@ -1,4 +1,0 @@
-package kr.co.popoolserver.inventory.domain.service;
-
-public interface InventoryService {
-}
