--- src/main/java/kr/co/popoolserver/inventory/domain/service/InventoryServiceImpl.java
@@ -1,4 +1,0 @@
-package kr.co.popoolserver.inventory.domain.service;
-
-public class InventoryServiceImpl {
-}
