--- src/main/java/kr/co/popoolserver/inventory/repository/InventoryRepository.java
@@ -1,4 +1,0 @@
-package kr.co.popoolserver.inventory.repository;
-
-public interface InventoryRepository {
-}
