--- src/main/java/kr/co/popoolserver/payment/controller/PaymentController.java
@@ -1,4 +1,0 @@
-package kr.co.popoolserver.payment.controller;
-
-public class PaymentController {
-}
