--- src/main/java/kr/co/popoolserver/payment/domain/PaymentDto.java
@@ -1,4 +1,0 @@
-package kr.co.popoolserver.payment.domain;
-
-public class PaymentDto {
-}
