--- src/main/java/kr/co/popoolserver/payment/domain/service/PaymentService.java
@@ -1,4 +1,0 @@
-package kr.co.popoolserver.payment.domain.service;
-
-public interface PaymentService {
-}
