--- src/main/java/kr/co/popoolserver/payment/domain/service/PaymentServiceImpl.java
@@ -1,4 +1,0 @@
-package kr.co.popoolserver.payment.domain.service;
-
-public class PaymentServiceImpl {
-}
