--- src/main/java/kr/co/popoolserver/payment/repository/PaymentRepository.java
@@ -1,4 +1,0 @@
-package kr.co.popoolserver.payment.repository;
-
-public interface PaymentRepository {
-}
