+++ src/main/java/kr/co/popoolserver/product/controller/InventoryController.java
@@ -1,0 +1,4 @@
+package kr.co.popoolserver.product.controller;
+
+public class InventoryController {
+}
