--- src/main/java/kr/co/popoolserver/product/domain/ProductDto.java
@@ -1,4 +1,0 @@
-package kr.co.popoolserver.product.domain;
-
-public class ProductDto {
-}
