+++ src/main/java/kr/co/popoolserver/product/domain/dto/PaymentDto.java
@@ -1,0 +1,4 @@
+package kr.co.popoolserver.product.domain.dto;
+
+public class PaymentDto {
+}
