+++ src/main/java/kr/co/popoolserver/product/domain/dto/ProductDto.java
@@ -1,0 +1,4 @@
+package kr.co.popoolserver.product.domain.dto;
+
+public class ProductDto {
+}
