+++ src/main/java/kr/co/popoolserver/product/domain/service/InventoryService.java
@@ -1,4 +1,4 @@
 package kr.co.popoolserver.product.domain.service;
 
-public class ProductServiceImpl {
+public class InventoryService {
 }
