+++ src/main/java/kr/co/popoolserver/product/domain/service/PaymentService.java
@@ -1,4 +1,4 @@
 package kr.co.popoolserver.product.domain.service;
 
-public class ProductService {
+public class PaymentService {
 }
