+++ src/main/java/kr/co/popoolserver/product/repository/InventoryRepository.java
@@ -1,0 +1,4 @@
+package kr.co.popoolserver.product.repository;
+
+public interface InventoryRepository {
+}
