+++ src/main/java/kr/co/popoolserver/product/repository/PaymentRepository.java
@@ -1,0 +1,4 @@
+package kr.co.popoolserver.product.repository;
+
+public interface PaymentRepository {
+}
